+++ tata-fne-back-end/src/main/java/com/bzdata/GestionCinema/attributionDroit/AttributionRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AttributionRepository extends JpaRepository<Attribution, Long> {
+public interface AttributionRepository extends JpaRepository<Attribution, Integer> {
     List<Attribution> findByUser_Id(int user_Id);
     List<Attribution> findByFunctionality_IdFunctionality(int functionality_Id);
     Optional<Attribution> findByUser_IdAndFunctionality_IdFunctionality(int user_Id, int functionality_Id);
